backend/sea_battle/requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
black==23.12.1
cffi==1.16.0
channels==4.0.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.0
daphne==4.0.0
Django==5.0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
install==1.3.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-dotenv==1.0.0
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
twisted==23.10.0
txaio==23.1.1
typing-extensions==4.9.0
tzdata==2023.4
zope.interface==6.1
